Update CategoryJoin.php
 convert from Integer to bigInteger

+++ app/Models/CategoryJoin.php
@@ -14,9 +14,9 @@ class CategoryJoin extends Model
     public $table="category_joins" ;
 
     protected $casts = [
-        'id' => 'integer',   
-        'post_id' => 'integer',   
-        'category_id' => 'integer',   
+        'id' => 'bigInteger',   
+        'post_id' => 'bigInteger',   
+        'category_id' => 'bigInteger',   
     ];
 
     protected $fillable = [
